thumbnails evm summit

+++ packages/tools/src/scripts/devconnect.ts
@@ -198,20 +198,21 @@ async function generateEventAssets(event: any) {
               }
 
               upload(thumbnailId, thumbnailFilePath, thumbnailType, folderId)
+            }
 
-              if (updateSessionThumbnails) {
-                const copyPath = join(publicEventFolder, thumbnailId)
-                copyFileSync(thumbnailFilePath, copyPath)
 
-                // update session file 
-                const sessionFilePath = join(dataSessionFolder, `${session.id}.json`)
-                if (existsSync(sessionFilePath)) {
-                  const sessionFile = JSON.parse(readFileSync(sessionFilePath, 'utf8'))
-                  writeFileSync(sessionFilePath, JSON.stringify({
-                    ...sessionFile,
-                    coverImage: `/sessions/${event.id}/${thumbnailId}`
-                  }, null, 2))
-                }
+            if (updateSessionThumbnails && fileExists(thumbnailFilePath)) {
+              const copyPath = join(publicEventFolder, thumbnailId)
+              copyFileSync(thumbnailFilePath, copyPath)
+
+              // update session file 
+              const sessionFilePath = join(dataSessionFolder, `${session.id}.json`)
+              if (existsSync(sessionFilePath)) {
+                const sessionFile = JSON.parse(readFileSync(sessionFilePath, 'utf8'))
+                writeFileSync(sessionFilePath, JSON.stringify({
+                  ...sessionFile,
+                  coverImage: `/sessions/${event.id}/${thumbnailId}`
+                }, null, 2))
               }
             }
           }
